+++ hiveengine/cli.py
@@ -4,7 +4,7 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 from builtins import bytes, int, str
-from beem import Steem
+from beem import Hive
 from beem.account import Account
 from beem.nodelist import NodeList
 from beem.utils import formatTimeString, addTzInfo
@@ -26,7 +26,7 @@ import io
 import argparse
 import re
 import six
-from beem.instance import set_shared_steem_instance, shared_steem_instance
+from beem.instance import set_shared_blockchain_instance, shared_blockchain_instance
 from hiveengine.version import version as __version__
 click.disable_unicode_literals_warning = True
 log = logging.getLogger(__name__)
@@ -89,9 +89,11 @@ def unlock_wallet(stm, password=None, allow_wif=True):
 @shell(prompt='hiveengine> ', intro='Starting hiveengine... (use help to list all commands)', chain=True)
 # click.group(chain=True)
 @click.option(
+    '--no-broadcast', '-d', is_flag=True, default=False, help="Do not broadcast")
+@click.option(
     '--verbose', '-v', default=3, help='Verbosity')
 @click.version_option(version=__version__)
-def cli(verbose):
+def cli(no_broadcast, verbose):
     # Logging
     log = logging.getLogger(__name__)
     verbosity = ["critical", "error", "warn", "info", "debug"][int(
@@ -104,6 +106,10 @@ def cli(verbose):
     ch.setFormatter(formatter)
     log.addHandler(ch)
     debug = verbose > 0
+    stm = Hive(
+        nobroadcast=no_broadcast,
+        )
+    set_shared_blockchain_instance(stm)
     pass
 
 
@@ -184,8 +190,8 @@ def info(objects):
             if stm is None:
                 nodelist = NodeList()
                 nodelist.update_nodes()
-                stm = Steem(node=nodelist.get_hive_nodes())                
-            wallet = Wallet(obj, steem_instance=stm)
+                stm = Hive(node=nodelist.get_hive_nodes())                
+            wallet = Wallet(obj, blockchain_instance=stm)
             t = PrettyTable(["id", "symbol", "balance", "stake", "pendingUnstake", "delegationsIn", "delegationsOut", "pendingUndelegations"])
             t.align = "l"
             for token in wallet:
@@ -242,22 +248,83 @@ def richlist(symbol, top):
 
 @cli.command()
 @click.argument('to', nargs=1)
-@click.argument('amount', nargs=1)
-@click.argument('token', nargs=1)
+@click.argument('amount', nargs=1, required=False)
+@click.argument('token', nargs=1, required=False)
 @click.argument('memo', nargs=1, required=False)
+@click.option('--memos', '-m', help="Can be used when all tokens should be send")
 @click.option('--account', '-a', help='Transfer from this account')
-def transfer(to, amount, token, memo, account):
+def transfer(to, amount, token, memo, memos, account):
     """Transfer a token"""
-    stm = shared_steem_instance()
+    stm = shared_blockchain_instance()
     if stm.rpc is not None:
         stm.rpc.rpcconnect()
+    if not stm.is_hive:
+        print("Please set a Hive node")
+        return
     if not account:
         account = stm.config["default_account"]
     if not bool(memo):
         memo = ''
+    if not bool(memos):
+        memos = ''
     if not unlock_wallet(stm):
         return
-    wallet = Wallet(account, steem_instance=stm)
+    
+    wallet = Wallet(account, blockchain_instance=stm)
+    if amount is None and token is None:
+        token = amount
+        amount = 0
+        tokens = Tokens()
+        
+        for t in wallet:
+            token = t["symbol"]
+            amount = float(t["balance"])
+            if amount == 0:
+                continue
+            token_obj = tokens.get_token(token)
+            market_info = token_obj.get_market_info()
+            if market_info is None:
+                print("transfer  %.8f %s to %s?" % (amount, token, to))
+            else:
+                last_price = float(market_info["lastPrice"])
+                highest_bid = float(market_info["highestBid"])
+                if highest_bid == 0:
+                    price = last_price
+                else:
+                    price = highest_bid
+                hive_amount = price*amount
+                print("transfer %.8f %s (value %.3f HIVE) to %s?" % (amount, token, hive_amount, to))
+            ret = input("continue [y/n]?")
+            if ret not in ["y", "yes"]:
+                continue
+            tx = wallet.transfer(to, amount, token, memos)
+            tx = json.dumps(tx, indent=4)
+            print(tx)
+        return
+    elif token is None:
+        token = amount
+        amount = 0
+        for t in wallet:
+            if t["symbol"] == token:
+                amount = float(t["balance"])
+        if amount == 0:
+            print("Amount of %s is 0" % token)
+            return
+        tokens = Tokens()
+        token_obj = tokens.get_token(token)   
+        market_info = token_obj.get_market_info()
+        last_price = float(market_info["lastPrice"])
+        highest_bid = float(market_info["highestBid"])
+        if highest_bid == 0:
+            price = last_price
+        else:
+            price = highest_bid
+        hive_amount = price*amount
+        print("transfer %.8f %s (value %.3f HIVE) to %s?" % (amount, token, hive_amount, to))
+        ret = input("continue [y/n]?")
+        if ret not in ["y", "yes"]:
+            return
+        memo = memos
     tx = wallet.transfer(to, amount, token, memo)
     tx = json.dumps(tx, indent=4)
     print(tx)
@@ -270,55 +337,183 @@ def transfer(to, amount, token, memo, account):
 @click.option('--account', '-a', help='Transfer from this account')
 def issue(to, amount, token, account):
     """Issue a token"""
-    stm = shared_steem_instance()
+    stm = shared_blockchain_instance()
     if stm.rpc is not None:
         stm.rpc.rpcconnect()
+    if not stm.is_hive:
+        print("Please set a Hive node")
+        return
     if not account:
         account = stm.config["default_account"]
     if not unlock_wallet(stm):
         return
-    wallet = Wallet(account, steem_instance=stm)
+    wallet = Wallet(account, blockchain_instance=stm)
     tx = wallet.issue(to, amount, token)
     tx = json.dumps(tx, indent=4)
     print(tx)
 
 
 @cli.command()
-@click.argument('amount', nargs=1)
-@click.argument('token', nargs=1)
-@click.option('--account', '-a', help='Take token from this account')
+@click.argument('amount', nargs=1, required=False)
+@click.argument('token', nargs=1, required=False)
+@click.option('--account', '-a', help='Stake token from this account')
 @click.option('--receiver', '-r', help='Stake to this account (default is sender account)')
 def stake(amount, token, account, receiver):
-    """stake a token"""
-    stm = shared_steem_instance()
+    """stake a token / all tokens"""
+    stm = shared_blockchain_instance()
     if stm.rpc is not None:
         stm.rpc.rpcconnect()
+    if not stm.is_hive:
+        print("Please set a Hive node")
+        return
     if not account:
         account = stm.config["default_account"]
     if not receiver:
         receiver = account
     if not unlock_wallet(stm):
         return
-    wallet = Wallet(account, steem_instance=stm)
+    wallet = Wallet(account, blockchain_instance=stm)
+    if amount is None and token is None:
+        token = amount
+        amount = 0
+        tokens = Tokens()
+        
+        for t in wallet:
+            token = t["symbol"]
+            amount = float(t["balance"])
+            if amount == 0:
+                continue
+            token_obj = tokens.get_token(token)
+            if not token_obj["stakingEnabled"]:
+                continue
+            market_info = token_obj.get_market_info()
+            if market_info is None:
+                print("stake %.8f %s?" % (amount, token))
+            else:
+                last_price = float(market_info["lastPrice"])
+                highest_bid = float(market_info["highestBid"])
+                if highest_bid == 0:
+                    price = last_price
+                else:
+                    price = highest_bid
+                hive_amount = price*amount
+                print("stake %.8f %s (value %.3f HIVE)?" % (amount, token, hive_amount))
+            ret = input("continue [y/n]?")
+            if ret not in ["y", "yes"]:
+                continue
+            tx = wallet.stake(amount, token, receiver=receiver)
+            tx = json.dumps(tx, indent=4)
+            print(tx)
+        return
+    elif token is None:
+        token = amount
+        amount = 0
+        for t in wallet:
+            if t["symbol"] == token:
+                amount = float(t["balance"])
+        if amount == 0:
+            print("Amount of %s is 0" % token)
+            return
+        tokens = Tokens()
+        token_obj = tokens.get_token(token)
+        if not token_obj["stakingEnabled"]:
+            print("%s is not stakable" % token)
+            return        
+        market_info = token_obj.get_market_info()
+        last_price = float(market_info["lastPrice"])
+        highest_bid = float(market_info["highestBid"])
+        if highest_bid == 0:
+            price = last_price
+        else:
+            price = highest_bid
+        hive_amount = price*amount
+        print("stake %.8f %s (value %.3f HIVE)?" % (amount, token, hive_amount))
+        ret = input("continue [y/n]?")
+        if ret not in ["y", "yes"]:
+            return
+
     tx = wallet.stake(amount, token, receiver=receiver)
     tx = json.dumps(tx, indent=4)
     print(tx)
 
 
 @cli.command()
-@click.argument('amount', nargs=1)
-@click.argument('token', nargs=1)
+@click.argument('amount', nargs=1, required=False)
+@click.argument('token', nargs=1, required=False)
 @click.option('--account', '-a', help='Transfer from this account')
 def unstake(amount, token, account):
-    """unstake a token"""
-    stm = shared_steem_instance()
+    """unstake a token / all tokens"""
+    stm = shared_blockchain_instance()
     if stm.rpc is not None:
         stm.rpc.rpcconnect()
+    if not stm.is_hive:
+        print("Please set a Hive node")
+        return
     if not account:
         account = stm.config["default_account"]
     if not unlock_wallet(stm):
         return
-    wallet = Wallet(account, steem_instance=stm)
+    wallet = Wallet(account, blockchain_instance=stm)
+    if amount is None and token is None:
+        token = amount
+        amount = 0
+        tokens = Tokens()
+        
+        for t in wallet:
+            token = t["symbol"]
+            amount = float(t["balance"])
+            stake = float(t["stake"])
+            if stake == 0:
+                continue
+            token_obj = tokens.get_token(token)
+            if not token_obj["stakingEnabled"]:
+                continue
+            market_info = token_obj.get_market_info()
+            if market_info is None:
+                print("unstake %.8f %s?" % (stake, token))
+            else:
+                last_price = float(market_info["lastPrice"])
+                highest_bid = float(market_info["highestBid"])
+                if highest_bid == 0:
+                    price = last_price
+                else:
+                    price = highest_bid
+                hive_amount = price*stake
+                print("unstake %.8f %s (value %.3f HIVE)?" % (stake, token, hive_amount))
+            ret = input("continue [y/n]?")
+            if ret not in ["y", "yes"]:
+                continue
+            tx = wallet.unstake(stake, token)
+            tx = json.dumps(tx, indent=4)
+            print(tx)
+        return
+    elif token is None:
+        token = amount
+        amount = 0
+        for t in wallet:
+            if t["symbol"] == token:
+                amount = float(t["stake"])
+        if amount == 0:
+            print("Staked Amount of %s is 0" % token)
+            return
+        tokens = Tokens()
+        token_obj = tokens.get_token(token)
+        if not token_obj["stakingEnabled"]:
+            print("%s is not stakable" % token)
+            return        
+        market_info = token_obj.get_market_info()
+        last_price = float(market_info["lastPrice"])
+        highest_bid = float(market_info["highestBid"])
+        if highest_bid == 0:
+            price = last_price
+        else:
+            price = highest_bid
+        hive_amount = price*amount
+        print("unstake %.8f %s (value %.3f HIVE)?" % (amount, token, hive_amount))
+        ret = input("continue [y/n]?")
+        if ret not in ["y", "yes"]:
+            return
+    
     tx = wallet.unstake(amount, token)
     tx = json.dumps(tx, indent=4)
     print(tx)
@@ -329,14 +524,17 @@ def unstake(amount, token, account):
 @click.option('--account', '-a', help='Transfer from this account')
 def cancel_unstake(account, trx_id):
     """unstake a token"""
-    stm = shared_steem_instance()
+    stm = shared_blockchain_instance()
     if stm.rpc is not None:
         stm.rpc.rpcconnect()
+    if not stm.is_hive:
+        print("Please set a Hive node")
+        return
     if not account:
         account = stm.config["default_account"]
     if not unlock_wallet(stm):
         return
-    wallet = Wallet(account, steem_instance=stm)
+    wallet = Wallet(account, blockchain_instance=stm)
     tx = wallet.cancel_unstake(trx_id)
     tx = json.dumps(tx, indent=4)
     print(tx)
@@ -347,14 +545,17 @@ def cancel_unstake(account, trx_id):
 @click.option('--account', '-a', help='withdraw from this account')
 def withdraw(amount, account):
     """Widthdraw SWAP.HIVE to account as HIVE."""
-    stm = shared_steem_instance()
+    stm = shared_blockchain_instance()
     if stm.rpc is not None:
         stm.rpc.rpcconnect()
+    if not stm.is_hive:
+        print("Please set a Hive node")
+        return
     if not account:
         account = stm.config["default_account"]
     if not unlock_wallet(stm):
         return
-    market = Market(steem_instance=stm)
+    market = Market(blockchain_instance=stm)
     tx = market.withdraw(account, amount)
     tx = json.dumps(tx, indent=4)
     print(tx)
@@ -365,14 +566,17 @@ def withdraw(amount, account):
 @click.option('--account', '-a', help='withdraw from this account')
 def deposit(amount, account):
     """Deposit HIVE to market in exchange for SWAP.HIVE."""
-    stm = shared_steem_instance()
+    stm = shared_blockchain_instance()
     if stm.rpc is not None:
         stm.rpc.rpcconnect()
+    if not stm.is_hive:
+        print("Please set a Hive node")
+        return
     if not account:
         account = stm.config["default_account"]
     if not unlock_wallet(stm):
         return
-    market = Market(steem_instance=stm)
+    market = Market(blockchain_instance=stm)
     tx = market.deposit(account, amount)
     tx = json.dumps(tx, indent=4)
     print(tx)
@@ -387,12 +591,15 @@ def buy(amount, token, price, account):
     """Put a buy-order for a token to the hive-engine market
 
     """
-    stm = shared_steem_instance()
+    stm = shared_blockchain_instance()
     if stm.rpc is not None:
         stm.rpc.rpcconnect()
+    if not stm.is_hive:
+        print("Please set a Hive node")
+        return
     if account is None:
         account = stm.config["default_account"]
-    market = Market(steem_instance=stm)
+    market = Market(blockchain_instance=stm)
     if not unlock_wallet(stm):
         return
     tx = market.buy(account, amount, token, price)
@@ -401,25 +608,140 @@ def buy(amount, token, price, account):
 
 
 @cli.command()
-@click.argument('amount', nargs=1)
-@click.argument('token', nargs=1)
-@click.argument('price', nargs=1)
+@click.argument('amount', nargs=1, required=False)
+@click.argument('token', nargs=1, required=False)
+@click.argument('price', nargs=1, required=False)
 @click.option('--account', '-a', help='Buy with this account (defaults to "default_account")')
 def sell(amount, token, price, account):
     """Put a sell-order for a token to the hive-engine market
 
     """
-    stm = shared_steem_instance()
+    stm = shared_blockchain_instance()
     if stm.rpc is not None:
         stm.rpc.rpcconnect()
+    if not stm.is_hive:
+        print("Please set a Hive node")
+        return
     if account is None:
         account = stm.config["default_account"]
-    market = Market(steem_instance=stm)
+
+    if amount is None and price is None and token is None:
+        if not unlock_wallet(stm):
+            return
+        token = amount
+        amount = 0
+        tokens = Tokens()
+        wallet = Wallet(account, blockchain_instance=stm)
+        market = Market(blockchain_instance=stm)
+        for t in wallet:
+            token = t["symbol"]
+            amount = float(t["balance"])
+            if amount == 0:
+                continue
+            token_obj = tokens.get_token(token)
+            market_info = token_obj.get_market_info()
+            if market_info is None:
+                continue
+            last_price = float(market_info["lastPrice"])
+            highest_bid = float(market_info["highestBid"])
+            if highest_bid == 0:
+                price = last_price
+            else:
+                price = highest_bid
+            hive_amount = price*amount
+            if hive_amount < 0.001:
+                continue
+            print("%s: using %.8f as price to sell %.8f %s for %.8f HIVE" % (token, price, amount, token, hive_amount))
+            ret = input("continue [y/n]?")
+            if ret not in ["y", "yes"]:
+                continue
+            tx = market.sell(account, amount, token, price)
+            tx = json.dumps(tx, indent=4)
+            print(tx)
+        return
+
+    elif price is None and token is None:
+        token = amount
+        amount = 0
+        wallet = Wallet(account, blockchain_instance=stm)
+        for t in wallet:
+            if t["symbol"] == token:
+                amount = float(t["balance"])
+        if amount == 0:
+            print("Amount of %s is 0" % token)
+            return
+        tokens = Tokens()
+        token_obj = tokens.get_token(token)
+        market_info = token_obj.get_market_info()
+        last_price = float(market_info["lastPrice"])
+        highest_bid = float(market_info["highestBid"])
+        if highest_bid == 0:
+            price = last_price
+        else:
+            price = highest_bid
+        hive_amount = price*amount
+        print("using %.8f as price to sell %.8f %s for %.8f HIVE" % (price, amount, token, hive_amount))
+        ret = input("continue [y/n]?")
+        if ret not in ["y", "yes"]:
+            return
+
+
+
+    market = Market(blockchain_instance=stm)
     if not unlock_wallet(stm):
         return
     tx = market.sell(account, amount, token, price)
     tx = json.dumps(tx, indent=4)
     print(tx)
+
+
+@cli.command()
+@click.option('--account', '-a', help='Buy with this account (defaults to "default_account")')
+def balance(account):
+    """Show token balance and value
+
+    """
+    stm = shared_blockchain_instance()
+    if stm.rpc is not None:
+        stm.rpc.rpcconnect()
+    if not stm.is_hive:
+        print("Please set a Hive node")
+        return
+    if account is None:
+        account = stm.config["default_account"]
+    token = ""
+    amount = 0
+    tokens = Tokens()
+    wallet = Wallet(account, blockchain_instance=stm)
+    table = PrettyTable(["symbol", "balance", "stake", "liquid HIVE", "staked HIVE"])
+    table.align = "l"    
+    sum_amount = 0
+    sum_staked = 0
+    for t in wallet:
+        token = t["symbol"]
+        amount = float(t["balance"])
+        stake = float(t["stake"])
+        token_obj = tokens.get_token(token)
+        market_info = token_obj.get_market_info()
+        if market_info is None:
+            continue
+        last_price = float(market_info["lastPrice"])
+        highest_bid = float(market_info["highestBid"])
+        if highest_bid == 0:
+            price = last_price
+        else:
+            price = highest_bid
+        balance_hive = price*amount
+        stake_hive = price*stake
+        sum_amount += balance_hive
+        sum_staked += stake_hive
+        if token_obj["stakingEnabled"]:
+            table.add_row([token, amount, stake, "%.3f" % balance_hive, "%.3f" % stake_hive])
+        else:
+            table.add_row([token, amount, stake, "%.3f" % balance_hive, "-"])
+    table.add_row(["-", "-", "-", "-", "-"])
+    table.add_row(["SUM", "", "", "%.3f" % sum_amount, "%.3f" % sum_staked])
+    print(table)
 
 
 @cli.command()
@@ -431,12 +753,15 @@ def cancel(order_type, order_id, account):
     
         order_type is either sell or buy
     """
-    stm = shared_steem_instance()
+    stm = shared_blockchain_instance()
     if stm.rpc is not None:
         stm.rpc.rpcconnect()
+    if not stm.is_hive:
+        print("Please set a Hive node")
+        return
     if account is None:
         account = stm.config["default_account"]
-    market = Market(steem_instance=stm)
+    market = Market(blockchain_instance=stm)
     if not unlock_wallet(stm):
         return
     tx = market.cancel(account, order_type, int(order_id))
@@ -451,10 +776,13 @@ def buybook(token, account):
     """Returns the buy book for the given token
 
     """
-    stm = shared_steem_instance()
+    stm = shared_blockchain_instance()
     if stm.rpc is not None:
         stm.rpc.rpcconnect()
-    market = Market(steem_instance=stm)
+    if not stm.is_hive:
+        print("Please set a Hive node")
+        return
+    market = Market(blockchain_instance=stm)
     buy_book = market.get_buy_book(token, account)
     sorted_buy_book = sorted(buy_book, key=lambda account: float(account["price"]), reverse=True)
     t = PrettyTable(["order_id", "account", "quantity", "price"])
@@ -470,10 +798,13 @@ def buybook(token, account):
 def sellbook(token, account):
     """Returns the sell book for the given token
     """
-    stm = shared_steem_instance()
+    stm = shared_blockchain_instance()
     if stm.rpc is not None:
         stm.rpc.rpcconnect()
-    market = Market(steem_instance=stm)
+    if not stm.is_hive:
+        print("Please set a Hive node")
+        return
+    market = Market(blockchain_instance=stm)
     sell_book = market.get_sell_book(token, account)
     sorted_sell_book = sorted(sell_book, key=lambda account: float(account["price"]), reverse=False)
     t = PrettyTable(["order_id", "account", "quantity", "price"])