+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+## 0.1.4
+* allow so transfer / stake / unstake / sell all tokens at once
+* When no amount is given by transfer / stake / unstake the token balance is used
+* new balance command
+
 ## 0.1.3
 * some bug-fixes